Refactor to rename func/dir name
